+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [3.0.2] IN PROGRESS
 * [UIPQB-279](https://folio-org.atlassian.net/browse/UIPQB-279) Add support for fields containing a valueSourceApi property without a source property
+* [UIPQB-280](https://folio-org.atlassian.net/browse/UIPQB-280) Show labels for single-value custom field dropdowns in user-friendly queries
 
 ## [3.0.1](https://github.com/folio-org/ui-plugin-query-builder/tree/v3.0.1) (2026-05-13)
 * [UIPQB-271](https://folio-org.atlassian.net/browse/UIPQB-271) Disambiguate duplicate language labels in UI